+++ cli/src/helpers.rs
@@ -26,7 +26,7 @@ use crate::{
         DepositOpts, ShowSolidoAuthoritiesOpts, ShowSolidoOpts, WithdrawOpts,
     },
     error::CliError,
-    get_signer,
+    get_signer_from_path,
 };
 use crate::{
     multisig::{get_multisig_program_address, propose_instruction, ProposeInstructionOutput},
@@ -106,7 +106,7 @@ pub fn command_create_solido(
     let lido_signer = {
         if opts.solido_key_path() != &PathBuf::default() {
             // If we've been given a solido private key, use it to create the solido instance.
-            get_signer(opts.solido_key_path().clone())?
+            get_signer_from_path(opts.solido_key_path().clone())?
         } else {
             // If not, use a random key
             Box::new(Keypair::new())